Fix a couple of dependency issues
The OS Perl was upgraded, so all the libraries needed to be upgraded. Turns
out a few of our our dependencies were still installed in the system Perl
paths on the old OS. With the upgrade they're gone, so update the build tasks
to add those now missing dependencies to local libs.