Switch from cron to systemd
Depesz asked for the switch. Conveniently Ansible has built in support
for systemd. So add service files for manager, consumer, api, and site
and update the deploy and restart playbooks to deploy, configure, and
restart the services. Leave the `_ctl` scripts in place for future
reference.

While at it, fix linting issues in the modified Ansible playbooks and
set `ansible_python_interpreter` for each host to suppress the
"discovered Python interpreter" warning.